requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
easyocr==1.7.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.0
gTTS==2.5.1
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.3.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pygame==2.5.2
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
sympy==1.12.1
tifffile==2024.6.18
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.40
ultralytics-thop==2.0.0
urllib3==2.2.2
Werkzeug==3.0.3
